Rename new-messages.ts to messages.ts
