+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/auto-model-codeml-queries.test.ts
@@ -4,7 +4,7 @@ import {
   DatabaseKind,
 } from "../../../../src/databases/local-databases";
 import { file } from "tmp-promise";
-import { QueryResultType } from "../../../../src/query-server/new-messages";
+import { QueryResultType } from "../../../../src/query-server/messages";
 import {
   generateCandidateFilterPack,
   runAutoModelQueries,
